+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## 0.3.2
 
-- Prevent caching of HTTP requests.
+- Implement `../PATH` and `./PATH` URLs relative to Cacao's program file.
+- Use MJPEG DataLen header if Content-Length doesn't exist.
+- Make serveFile send 404's and directory index based on the filesystem contents.
+- Send no-cache headers from serveHttp.
 
 ## 0.3.1